7. There are only 3 colors, 10 digits, and 7 notes; it's what we do with them that's important  (automated command)

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -6,7 +6,7 @@
  * Return: the value of the last digit
  *
  */
-int print_last_digit(int i)
+int print_last_digit(long long int i)
 {
 	int x;
